fix: get_brain_details API to include user_id parameter (#2242)
This pull request fixes the `get_brain_details` API by adding a
`user_id` parameter. Previously, the API did not include the `user_id`
parameter, which caused issues when retrieving brain details. This fix
ensures that the `user_id` is passed correctly to the API, allowing for
accurate retrieval of brain details.