docs: Add reranking configuration guide (#2489)
This pull request adds a new guide to the documentation that explains
how to configure reranking in Quivr. The guide covers the steps to
configure the reranking service using the FlashrankRerank library and
the Cohere ReRanker. It also provides information on the required
environment variables and the default behavior of the Ranker.