feat: Fix bug in generate_source function
The generate_source function in the rag_service/utils.py file had a bug that caused an error when generating the source URL for a document. This bug has been fixed by adding exception handling around the code that generates the URL. Now, if an error occurs during URL generation, the function will log an error message and continue to the next document.