feat: Add assistant_id and settings to CreateTask DTO and Task entity
The code changes in `inputs.py` and `task_entity.py` add the `assistant_id` field of type int and the `settings` field of type dict to the `CreateTask` DTO and `Task` entity respectively. This allows for capturing the assistant ID and associated settings when creating a new task. This feature enhances the functionality of the assistant module.