feat: Add file name to knowledge properties in crawl_endpoint (#3346)
The `crawl_endpoint` function in `crawl_routes.py` now includes the
`file_name` property in the `knowledge_to_add` object. This change
ensures that the file name is properly associated with the crawled
knowledge, improving the accuracy and organization of the data.

Co-authored-by: Stan Girard <girard.stanislas@gmail.com>

# Description

Please include a summary of the changes and the related issue. Please
also include relevant motivation and context.

## Checklist before requesting a review

Please delete options that are not relevant.

- [ ] My code follows the style guidelines of this project
- [ ] I have performed a self-review of my code
- [ ] I have commented hard-to-understand areas
- [ ] I have ideally added tests that prove my fix is effective or that
my feature works
- [ ] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged

## Screenshots (if appropriate):